split musician network into component + per-view renderer modules
MusicianNetworkComponent keeps the state machine and chrome (tabs,
slider sync, fullscreen, show-names toggle, tooltip plumbing + HTML
builders) and dispatches render() through a plain ctx object to three
new dumb renderer modules: networkGraphRenderer (force-directed graph),
networkMatrixRenderer (adjacency matrix), networkChordRenderer (chord
diagram). chordLabelVisibility moves with the chord renderer and is
re-exported from musicianNetworkComponent so existing test import
paths keep working. Mechanical move — no behavior change.
