+++ src/musicianNetworkComponent.js
@@ -1,18 +1,18 @@
 import * as d3 from "d3";
-import { getCssColor, getPartColor } from './config.js';
 import { escapeHtml } from './escapeHtml.js';
 import { tooltip } from './tooltip.js';
 import { MAX_DESIGN_WIDTH, isMobileWidth, isTouchPrimary } from './breakpoints.js';
 import {
     buildNetworkData,
     disambiguateLabels,
-    computeNodeCounts,
     defaultMinPiecesForGraph,
     computePartBreakdownPerMusician,
-    predominantPart,
     computeSliderSync,
     PART_ORDER,
 } from './dataProcessor.js';
+import { renderGraph } from './networkGraphRenderer.js';
+import { renderMatrix } from './networkMatrixRenderer.js';
+import { renderChord, chordLabelVisibility } from './networkChordRenderer.js';
 
 // Tabbed view of the musician co-occurrence network: a force-directed
 // graph and an adjacency matrix over the same top-N node set.
@@ -20,32 +20,17 @@ import {
 // Both views consume the cached `_state` recomputed in render(), so
 // switching tabs is free. The user (spreadsheet owner) is excluded
 // implicitly via the userKey passed in by DashboardComponent.
+//
+// This module owns the state machine + chrome (tabs, slider, fullscreen,
+// show-names toggle, tooltip plumbing); the actual drawing lives in the
+// per-view renderer modules (networkGraphRenderer / networkMatrixRenderer /
+// networkChordRenderer), which consume a plain ctx object built here.
 
 const MIN_EDGE_WEIGHT = 2;
 
-// Greedy de-overlap for chord arc labels: walk groups around the circle,
-// tracking the tangential right edge of the last shown label, and hide any
-// label whose left edge would intrude on it. Effect on dense arcs: roughly
-// alternating labels (show one, skip the next), which beats a wall of
-// overlapping text; the tooltip still has the full name for hidden arcs.
-// Pure — exported for tests. `groups` are d3.chord() groups ({startAngle,
-// endAngle}), fontFor(group) → px, labelRadius the label ring radius.
-export function chordLabelVisibility(groups, fontFor, labelRadius) {
-    const visible = new Array(groups.length).fill(true);
-    let lastShownEnd = -Infinity;
-    groups.forEach((d, i) => {
-        const font = fontFor(d);
-        const mid = (d.startAngle + d.endAngle) / 2;
-        const halfAngular = (font / 2) / labelRadius;
-        if (mid - halfAngular >= lastShownEnd) {
-            visible[i] = true;
-            lastShownEnd = mid + halfAngular;
-        } else {
-            visible[i] = false;
-        }
-    });
-    return visible;
-}
+// chordLabelVisibility lives with the chord renderer; re-exported here so
+// tests (and any other consumer) keep their existing import path.
+export { chordLabelVisibility };
 
 function sizing(width) {
     const mobile = isMobileWidth(width);
@@ -264,16 +249,34 @@ export class MusicianNetworkComponent {
             return;
         }
 
+        // Plain context object consumed by the renderer modules: cached
+        // network state + sizing knobs + selection/toggles, plus callbacks
+        // into the component's tooltip plumbing so the renderers stay dumb.
+        const ctx = {
+            width,
+            s,
+            state: this._state,
+            selected: this.getSelectedMusician ? this.getSelectedMusician() : null,
+            showNames: this.showNames,
+            minEdgeWeight: MIN_EDGE_WEIGHT,
+            attachTooltip: (selection, getHtml) => this._attachTooltip(selection, getHtml),
+            attachHoverTooltip: (selection, getHtml) => this._attachHoverTooltip(selection, getHtml),
+            attachClickToggle: (selection, getName) => this._attachClickToggle(selection, getName),
+            nodeTooltipHtml: (n) => this._nodeTooltipHtml(n),
+            edgeTooltipHtml: (e) => this._edgeTooltipHtml(e),
+            cellTooltipHtml: (c) => this._cellTooltipHtml(c),
+        };
+
         const n = this._state.nodes.length;
         if (this.activeView === 'graph') {
             caption.text(`${n} co-player${n === 1 ? '' : 's'} · edges shown when ≥ ${MIN_EDGE_WEIGHT} shared pieces`);
-            this._renderGraph(width, s);
+            renderGraph(ctx);
         } else if (this.activeView === 'matrix') {
             caption.text(`${n} co-player${n === 1 ? '' : 's'} · diagonal omitted · cell shade = pieces played together`);
-            this._renderMatrix(width, s);
+            renderMatrix(ctx);
         } else {
             caption.text(`${n} co-player${n === 1 ? '' : 's'} · grouped by predominant instrument · ribbon = pieces played together`);
-            this._renderChord(width, s);
+            renderChord(ctx);
         }
     }
 
@@ -298,197 +301,9 @@ export class MusicianNetworkComponent {
         this._state = { nodes, edges, labels, maxNodeCount, maxEdgeWeight };
     }
 
-    // ---------------- Graph ----------------
-
-    _renderGraph(width, s) {
-        const root = d3.select('#dashboardMusicianNetworkGraph');
-        const height = s.graphHeight;
-        const { nodes: stateNodes, edges: stateEdges, labels, maxNodeCount, maxEdgeWeight } = this._state;
-
-        // Local mutable copies — d3-force rewrites source/target to node refs
-        // and mutates x/y, vx/vy on the node objects.
-        const nodes = stateNodes.map(n => ({ ...n, label: labels.get(n.name) }));
-        const edges = stateEdges.map(e => ({ ...e }));
-
-        const radiusScale = d3.scaleSqrt()
-            .domain([1, Math.max(1, maxNodeCount)])
-            .range(s.nodeRadiusRange);
-        const widthScale = d3.scaleSqrt()
-            .domain([MIN_EDGE_WEIGHT, Math.max(MIN_EDGE_WEIGHT, maxEdgeWeight)])
-            .range(s.edgeWidthRange);
-
-        const nodeRadius = n => radiusScale(n.count);
-        const labelHalfWidth = n => (n.label.length * s.labelFont * 0.3);
-
-        const sim = d3.forceSimulation(nodes)
-            .force('link', d3.forceLink(edges)
-                .id(d => d.name)
-                .distance(s.linkDistance)
-                .strength(d => 0.05 + 0.25 * d.weight / Math.max(1, maxEdgeWeight)))
-            .force('charge', d3.forceManyBody().strength(s.chargeStrength))
-            .force('center', d3.forceCenter(width / 2, height / 2))
-            .force('x', d3.forceX(width / 2).strength(0.04))
-            .force('y', d3.forceY(height / 2).strength(0.07))
-            .force('collide', d3.forceCollide(d => nodeRadius(d) + labelHalfWidth(d) + 4));
-
-        sim.stop();
-        for (let i = 0; i < 300; i++) sim.tick();
-
-        // Clamp positions inside the SVG so labels never get clipped.
-        const margin = 8;
-        nodes.forEach(n => {
-            const r = nodeRadius(n);
-            const lw = labelHalfWidth(n) * 2 + s.labelDx;
-            n.x = Math.max(r + margin, Math.min(width - r - lw - margin, n.x));
-            n.y = Math.max(r + margin, Math.min(height - r - margin, n.y));
-        });
-
-        // Pick label side: prefer right; flip to left if label would overshoot.
-        nodes.forEach(n => {
-            const r = nodeRadius(n);
-            const lw = labelHalfWidth(n) * 2;
-            n.labelOnRight = (n.x + r + s.labelDx + lw) <= (width - margin);
-        });
-
-        const edgeColor = getCssColor('--color-border-strong') || getCssColor('--color-text-secondary');
-        const selectedStroke = getCssColor('--color-text-dark') || getCssColor('--color-text-primary');
-        const sliceStroke = getCssColor('--color-bg-primary');
-        const otherFill = getCssColor('--color-part-fallback');
-        const selected = this.getSelectedMusician ? this.getSelectedMusician() : null;
-        const isEdgeIncident = e => selected && (
-            (e.source.name ?? e.source) === selected || (e.target.name ?? e.target) === selected
-        );
-
-        // Pie-arc helpers for the node breakdown. Each non-zero part bucket
-        // becomes one slice. The slices add up to the node's total piece
-        // count so the pie fills the full node circle.
-        const pieGen = d3.pie().value(d => d.count).sort(null);
-        const slicesFor = (n) => {
-            const parts = n.parts ?? { V1: 0, V2: 0, VA: 0, VC: 0, OTHER: 0 };
-            const entries = PART_ORDER
-                .map(part => ({ part, count: parts[part] ?? 0 }))
-                .filter(p => p.count > 0);
-            // Fallback for the unlikely empty-vector case: paint as a single
-            // accent-colored disc so the node is still visible.
-            if (entries.length === 0) entries.push({ part: null, count: 1 });
-            return pieGen(entries);
-        };
-        const sliceFill = (part) => {
-            if (part === null) return getCssColor('--color-accent');
-            if (part === 'OTHER') return otherFill;
-            return getPartColor(part);
-        };
-
-        // Build SVG fresh each render — node count is small and re-laying out
-        // is the dominant cost; redraw is negligible.
-        root.selectAll('*').remove();
-        const svg = root.append('svg')
-            .attr('width', width)
-            .attr('height', height)
-            .attr('viewBox', `0 0 ${width} ${height}`)
-            .style('display', 'block');
-
-        // Defs for the Voronoi hit layer's per-node clip circles. The hit
-        // group itself is appended below, AFTER the edges, so that taps
-        // inside a node's clip circle hit the node and not an edge passing
-        // through (edges can still catch taps outside any clip circle).
-        const delaunay = d3.Delaunay.from(nodes, d => d.x, d => d.y);
-        const voronoi = delaunay.voronoi([0, 0, width, height]);
-        const clipIdFor = i => `network-hit-clip-${i}`;
-        const defsSel = svg.append('defs');
-        defsSel.selectAll('clipPath')
-            .data(nodes)
-            .join('clipPath')
-            .attr('id', (d, i) => clipIdFor(i))
-            .append('circle')
-            .attr('cx', d => d.x)
-            .attr('cy', d => d.y)
-            .attr('r', s.nodeHitClipRadius);
-
-        const linkSel = svg.append('g').attr('class', 'network-edges')
-            .selectAll('line')
-            .data(edges)
-            .join('line')
-            .attr('class', 'network-edge')
-            .attr('x1', d => d.source.x)
-            .attr('y1', d => d.source.y)
-            .attr('x2', d => d.target.x)
-            .attr('y2', d => d.target.y)
-            .attr('stroke', edgeColor)
-            .attr('stroke-width', d => widthScale(d.weight))
-            .attr('stroke-opacity', d => {
-                const base = 0.25 + 0.45 * (d.weight / Math.max(1, maxEdgeWeight));
-                if (!selected) return base;
-                return isEdgeIncident(d) ? base : 0.08;
-            });
-        this._attachTooltip(linkSel, (event, d) => this._edgeTooltipHtml(d));
-
-        // Voronoi hit layer above the edges — each node's hit region is its
-        // Voronoi cell clipped to a circle around the node center, so taps
-        // anywhere in that circle hit the node, even when an edge passes
-        // through. Outside any clip circle this layer is transparent to
-        // events so edges below remain clickable for their tooltip.
-        const hitSel = svg.append('g').attr('class', 'network-hit')
-            .selectAll('path')
-            .data(nodes)
-            .join('path')
-            .attr('d', (d, i) => voronoi.renderCell(i))
-            .attr('clip-path', (d, i) => `url(#${clipIdFor(i)})`)
-            .attr('fill', 'transparent')
-            .attr('pointer-events', 'all');
-        this._attachClickToggle(hitSel, d => d.name);
-        this._attachHoverTooltip(hitSel, (event, d) => this._nodeTooltipHtml(d));
-
-        // Each node is a <g> at (x, y) with pie slices + a selection-outline
-        // circle + a transparent overlay that absorbs clicks/hovers (so a
-        // single handler set serves the whole node regardless of which slice
-        // the user lands on).
-        const nodeG = svg.append('g').attr('class', 'network-nodes')
-            .selectAll('g.network-node')
-            .data(nodes, d => d.name)
-            .join('g')
-            .attr('class', 'network-node')
-            .attr('transform', d => `translate(${d.x}, ${d.y})`)
-            .attr('opacity', d => !selected || d.name === selected ? 1 : 0.35);
-
-        nodeG.each(function (d) {
-            const r = nodeRadius(d);
-            const arcGen = d3.arc().innerRadius(0).outerRadius(r);
-            const g = d3.select(this);
-            g.selectAll('path.network-slice')
-                .data(slicesFor(d), s => s.data.part ?? 'fallback')
-                .join('path')
-                .attr('class', 'network-slice')
-                .attr('d', arcGen)
-                .attr('fill', s => sliceFill(s.data.part))
-                .attr('stroke', sliceStroke)
-                .attr('stroke-width', 0.5)
-                .attr('pointer-events', 'none');
-            g.selectAll('circle.network-node-outline')
-                .data([d])
-                .join('circle')
-                .attr('class', 'network-node-outline')
-                .attr('r', r)
-                .attr('fill', 'none')
-                .attr('stroke', d.name === selected ? selectedStroke : 'none')
-                .attr('stroke-width', d.name === selected ? 2 : 0)
-                .attr('pointer-events', 'none');
-        });
-
-        svg.append('g').attr('class', 'network-labels')
-            .selectAll('text')
-            .data(nodes)
-            .join('text')
-            .attr('class', 'network-label')
-            .attr('font-size', s.labelFont)
-            .attr('x', d => d.labelOnRight ? d.x + nodeRadius(d) + s.labelDx : d.x - nodeRadius(d) - s.labelDx)
-            .attr('y', d => d.y)
-            .attr('dy', '0.32em')
-            .attr('text-anchor', d => d.labelOnRight ? 'start' : 'end')
-            .attr('opacity', d => !selected || d.name === selected ? 1 : 0.35)
-            .attr('font-weight', d => d.name === selected ? 'bold' : 'normal')
-            .text(d => this.showNames ? d.label : '');
-    }
+    // ---------------- Tooltip HTML builders ----------------
+    // Kept on the prototype (not in the renderer modules) — they read
+    // this._state and are pinned by test/escapeHtml.test.mjs.
 
     _nodeTooltipHtml(n) {
         const { edges } = this._state;
@@ -511,357 +326,11 @@ export class MusicianNetworkComponent {
         return `<h4>${escapeHtml(a)} · ${escapeHtml(b)}</h4><ul><li>${e.weight} pieces together</li></ul>`;
     }
 
-    // ---------------- Matrix ----------------
-
-    _renderMatrix(containerWidth, s) {
-        const root = d3.select('#dashboardMusicianNetworkMatrix');
-        const { nodes, edges, labels, maxEdgeWeight } = this._state;
-        const n = nodes.length;
-        // Captured locally so the d3 .each callbacks (where `this` is the
-        // DOM element) can still gate name rendering off the toggle.
-        const showNames = this.showNames;
-
-        const labelGutter = s.matrixLabelGutter;
-        const availableForCells = containerWidth - labelGutter - 4;
-        // Clamped both ways: shrinks to matrixCellMin (horizontal scroll
-        // when the matrix exceeds container width), and capped at
-        // matrixCellMax so a sparse matrix (after selecting a musician
-        // and only ~10 cells remain) doesn't blow up into giant squares.
-        const cellSize = Math.max(
-            s.matrixCellMin,
-            Math.min(s.matrixCellMax, Math.floor(availableForCells / n))
-        );
-        const gridSize = cellSize * n;
-        const svgWidth = labelGutter + gridSize;
-        const svgHeight = labelGutter + gridSize;
-
-        // Build a {name → index} map and a sparse {key → weight} map.
-        const index = new Map(nodes.map((node, i) => [node.name, i]));
-        const weights = new Map();
-        edges.forEach(e => {
-            const a = e.source.name ?? e.source;
-            const b = e.target.name ?? e.target;
-            const i = index.get(a);
-            const j = index.get(b);
-            if (i === undefined || j === undefined) return;
-            weights.set(`${i},${j}`, e.weight);
-            weights.set(`${j},${i}`, e.weight);
-        });
-
-        // Quantile-based color: each cell's shade is set by its rank among
-        // all positive co-occurrence counts, not by absolute magnitude. The
-        // distribution is heavily right-skewed (a few very strong pairs, a
-        // long tail of weak ones), so a linear scale compresses the tail
-        // into indistinguishable near-zero greens. Ranks spread the cells
-        // evenly across the ramp. The ramp itself starts at 0.15 (a clearly
-        // green tint, not near-white) so the lightest filled cell stays
-        // distinguishable from empty cells. Tooltip gives exact counts.
-        const positiveWeights = edges.map(e => e.weight);
-        const ramp = t => d3.interpolateGreens(0.15 + 0.85 * t);
-        const color = positiveWeights.length > 0
-            ? d3.scaleSequentialQuantile(ramp).domain(positiveWeights)
-            : d3.scaleSequential(ramp).domain([0, 1]);
-        const emptyFill = getCssColor('--color-bg-stripe') || getCssColor('--color-bg-secondary');
-        const labelColor = getCssColor('--color-text-secondary');
-        const selectedLabelColor = getCssColor('--color-text-primary');
-        const selected = this.getSelectedMusician ? this.getSelectedMusician() : null;
-
-        root.selectAll('*').remove();
-        const svg = root.append('svg')
-            .attr('width', svgWidth)
-            .attr('height', svgHeight)
-            .attr('viewBox', `0 0 ${svgWidth} ${svgHeight}`)
-            .style('display', 'block');
-
-        // Row labels (left gutter), aligned to row centers.
-        const rowLabelSel = svg.append('g').attr('class', 'matrix-row-labels')
-            .selectAll('text')
-            .data(nodes)
-            .join('text')
-            .attr('class', 'matrix-label')
-            .attr('x', labelGutter - 6)
-            .attr('y', (d, i) => labelGutter + i * cellSize + cellSize / 2)
-            .attr('dy', '0.32em')
-            .attr('text-anchor', 'end')
-            .attr('font-size', s.matrixLabelFont)
-            .attr('fill', d => d.name === selected ? selectedLabelColor : labelColor)
-            .attr('font-weight', d => d.name === selected ? 'bold' : 'normal')
-            .attr('opacity', d => !selected || d.name === selected ? 1 : 0.5)
-            .text(d => showNames ? this._truncate(labels.get(d.name), labelGutter - 10, s.matrixLabelFont) : '')
-            .each(function (d) {
-                d3.select(this).append('title').text(showNames ? d.name : '');
-            });
-        this._attachClickToggle(rowLabelSel, d => d.name);
-        this._attachHoverTooltip(rowLabelSel, (event, d) => this._nodeTooltipHtml(d));
-
-        // Column labels (top gutter), rotated -90° so they read bottom-up
-        // with the head tilted left. Perfectly vertical (vs. tilted) avoids
-        // the rightmost label running off the gutter at the matrix edge.
-        // dominant-baseline=central puts the text's vertical center on the
-        // column center; text-anchor=start anchors the first letter just
-        // above the grid so reading direction is bottom→top.
-        const colLabelSel = svg.append('g').attr('class', 'matrix-col-labels')
-            .selectAll('text')
-            .data(nodes)
-            .join('text')
-            .attr('class', 'matrix-label')
-            .attr('transform', (d, i) => {
-                const cx = labelGutter + i * cellSize + cellSize / 2;
-                const cy = labelGutter - 6;
-                return `translate(${cx}, ${cy}) rotate(-90)`;
-            })
-            .attr('text-anchor', 'start')
-            .attr('dominant-baseline', 'central')
-            .attr('font-size', s.matrixLabelFont)
-            .attr('fill', d => d.name === selected ? selectedLabelColor : labelColor)
-            .attr('font-weight', d => d.name === selected ? 'bold' : 'normal')
-            .attr('opacity', d => !selected || d.name === selected ? 1 : 0.5)
-            .text(d => showNames ? this._truncate(labels.get(d.name), labelGutter - 10, s.matrixLabelFont) : '')
-            .each(function (d) {
-                d3.select(this).append('title').text(showNames ? d.name : '');
-            });
-        this._attachClickToggle(colLabelSel, d => d.name);
-        this._attachHoverTooltip(colLabelSel, (event, d) => this._nodeTooltipHtml(d));
-
-        // Cells. Build a flat array of {i, j, a, b, weight}; skip diagonal.
-        const cells = [];
-        for (let i = 0; i < n; i++) {
-            for (let j = 0; j < n; j++) {
-                if (i === j) continue;
-                cells.push({
-                    i, j,
-                    a: nodes[i].name,
-                    b: nodes[j].name,
-                    weight: weights.get(`${i},${j}`) ?? 0,
-                });
-            }
-        }
-
-        const cellSel = svg.append('g').attr('class', 'matrix-cells')
-            .selectAll('rect')
-            .data(cells)
-            .join('rect')
-            .attr('class', 'matrix-cell')
-            .attr('x', d => labelGutter + d.j * cellSize)
-            .attr('y', d => labelGutter + d.i * cellSize)
-            .attr('width', cellSize)
-            .attr('height', cellSize)
-            .attr('fill', d => d.weight === 0 ? emptyFill : color(d.weight));
-        this._attachTooltip(cellSel, (event, d) => this._cellTooltipHtml(d));
-
-        // Diagonal placeholder rects (no fill, no tooltip).
-        svg.append('g').attr('class', 'matrix-diagonal')
-            .selectAll('rect')
-            .data(nodes)
-            .join('rect')
-            .attr('class', 'matrix-cell-empty')
-            .attr('x', (d, i) => labelGutter + i * cellSize)
-            .attr('y', (d, i) => labelGutter + i * cellSize)
-            .attr('width', cellSize)
-            .attr('height', cellSize)
-            .attr('fill', 'transparent');
-    }
-
     _cellTooltipHtml(c) {
         if (c.weight === 0) {
             return `<h4>${escapeHtml(c.a)} · ${escapeHtml(c.b)}</h4><ul><li>No pieces together</li></ul>`;
         }
         return `<h4>${escapeHtml(c.a)} · ${escapeHtml(c.b)}</h4><ul><li>${c.weight} piece${c.weight === 1 ? '' : 's'} together</li></ul>`;
-    }
-
-    // ---------------- Chord ----------------
-
-    _renderChord(containerWidth, s) {
-        const root = d3.select('#dashboardMusicianNetworkChord');
-        const { nodes, edges, labels } = this._state;
-
-        // Group musicians by predominant instrument (V1 → V2 → VA → VC → OTHER),
-        // sorted by piece count desc within each block. The chord layout then
-        // arranges them in this order around the circle so each instrument
-        // family occupies a contiguous arc segment.
-        const order = PART_ORDER;
-        const ordered = nodes.slice().sort((a, b) => {
-            const pa = predominantPart(a.parts) ?? 'OTHER';
-            const pb = predominantPart(b.parts) ?? 'OTHER';
-            const oa = order.indexOf(pa);
-            const ob = order.indexOf(pb);
-            if (oa !== ob) return oa - ob;
-            return b.count - a.count;
-        });
-
-        const N = ordered.length;
-        const indexOf = new Map(ordered.map((n, i) => [n.name, i]));
-
-        // Build symmetric co-occurrence matrix.
-        const matrix = Array.from({ length: N }, () => new Array(N).fill(0));
-        edges.forEach(e => {
-            const a = e.source.name ?? e.source;
-            const b = e.target.name ?? e.target;
-            const i = indexOf.get(a);
-            const j = indexOf.get(b);
-            if (i === undefined || j === undefined) return;
-            matrix[i][j] = e.weight;
-            matrix[j][i] = e.weight;
-        });
-
-        // Square container, capped by both the section width and the per-
-        // breakpoint design size. The chord diagram inscribes a circle into it.
-        const diameter = Math.min(containerWidth, s.chordDiameter);
-        const outerRadius = (diameter / 2) - s.chordLabelPad;
-        const innerRadius = outerRadius - s.chordArcThickness;
-
-        const chord = d3.chord()
-            .padAngle(0.015)
-            .sortGroups(null)
-            .sortSubgroups(null);
-        const layout = chord(matrix);
-        const arcGen = d3.arc().innerRadius(innerRadius).outerRadius(outerRadius);
-        const ribbonGen = d3.ribbon().radius(innerRadius);
-
-        const selected = this.getSelectedMusician ? this.getSelectedMusician() : null;
-        const otherFill = getCssColor('--color-part-fallback');
-        const selectedStroke = getCssColor('--color-text-dark') || getCssColor('--color-text-primary');
-
-        const arcFill = (i) => {
-            const part = predominantPart(ordered[i].parts) ?? 'OTHER';
-            return part === 'OTHER' ? otherFill : getPartColor(part);
-        };
-        // Blended ribbon color (per request). If it ends up muddy on real data
-        // we can swap to a neutral gray here without touching the rest.
-        const ribbonFill = (i, j) => d3.interpolateRgb(arcFill(i), arcFill(j))(0.5);
-
-        root.selectAll('*').remove();
-        const svg = root.append('svg')
-            .attr('width', diameter)
-            .attr('height', diameter)
-            .attr('viewBox', `${-diameter / 2} ${-diameter / 2} ${diameter} ${diameter}`)
-            .style('display', 'block');
-
-        // Ribbons (chords) — drawn first so arcs sit on top.
-        const chordSel = svg.append('g').attr('class', 'network-chords')
-            .selectAll('path')
-            .data(layout, d => `${ordered[d.source.index].name}::${ordered[d.target.index].name}`)
-            .join('path')
-            .attr('class', 'network-chord')
-            .attr('d', ribbonGen)
-            .attr('fill', d => ribbonFill(d.source.index, d.target.index))
-            .attr('fill-opacity', d => {
-                if (!selected) return 0.45;
-                const a = ordered[d.source.index].name;
-                const b = ordered[d.target.index].name;
-                return (a === selected || b === selected) ? 0.75 : 0.04;
-            });
-        this._attachTooltip(chordSel, (event, d) => {
-            const a = ordered[d.source.index].name;
-            const b = ordered[d.target.index].name;
-            const w = matrix[d.source.index][d.target.index];
-            return `<h4>${escapeHtml(a)} · ${escapeHtml(b)}</h4><ul><li>${w} piece${w === 1 ? '' : 's'} together</li></ul>`;
-        });
-
-        // Outer arcs + labels.
-        const arcG = svg.append('g').attr('class', 'network-arcs')
-            .selectAll('g.network-arc-group')
-            .data(layout.groups, d => ordered[d.index].name)
-            .join('g')
-            .attr('class', 'network-arc-group');
-
-        const arcPathSel = arcG.append('path')
-            .attr('class', 'network-arc')
-            .attr('d', arcGen)
-            .attr('fill', d => arcFill(d.index))
-            .attr('opacity', d => {
-                const name = ordered[d.index].name;
-                return !selected || name === selected ? 1 : 0.35;
-            })
-            .attr('stroke', d => ordered[d.index].name === selected ? selectedStroke : 'none')
-            .attr('stroke-width', d => ordered[d.index].name === selected ? 2 : 0);
-
-        // Transparent hit arc, sized to the visible arc (chordHitPad = 0) on
-        // desktop and padded outward on touch. Sits on top of the visible arc
-        // so it captures pointer events; visible arc has pointer-events:none
-        // so it doesn't double-fire.
-        arcPathSel.attr('pointer-events', 'none');
-        const arcHitGen = d3.arc()
-            .innerRadius(Math.max(0, innerRadius - s.chordHitPad))
-            .outerRadius(outerRadius + s.chordHitPad);
-        const arcHitSel = arcG.append('path')
-            .attr('class', 'network-arc-hit')
-            .attr('d', arcHitGen)
-            .attr('fill', 'transparent')
-            .attr('pointer-events', 'all');
-        this._attachClickToggle(arcHitSel, d => ordered[d.index].name);
-        this._attachHoverTooltip(arcHitSel, (event, d) => this._nodeTooltipHtml(ordered[d.index]));
-
-        // Radial labels just outside the arcs. The conditional rotate(180)
-        // flips text on the left half of the circle so it always reads
-        // outward-to-inward rather than upside-down.
-        //
-        // Font size: cap each label at the per-arc tangential budget
-        // (arcAngular × labelRadius), clamped to a floor — adjacent labels
-        // compete tangentially, not radially. If even the floor doesn't
-        // fit, the greedy visibility pass below alternately hides labels
-        // so survivors stay legible.
-        const labelRadius = outerRadius + 6;
-        const MIN_LABEL_FONT = 7;
-        const fontFor = (d) => {
-            const arcAngular = d.endAngle - d.startAngle;
-            return Math.max(MIN_LABEL_FONT, Math.min(s.chordLabelFont, arcAngular * labelRadius));
-        };
-
-        const visible = chordLabelVisibility(layout.groups, fontFor, labelRadius);
-
-        arcG.append('text')
-            .attr('class', 'network-arc-label')
-            .attr('data-arc-index', (d, i) => i)
-            .attr('font-size', fontFor)
-            .attr('dy', '0.32em')
-            .attr('transform', d => {
-                const angleDeg = (d.startAngle + d.endAngle) / 2 * 180 / Math.PI - 90;
-                const flip = angleDeg > 90;
-                return `rotate(${angleDeg}) translate(${labelRadius})${flip ? ' rotate(180)' : ''}`;
-            })
-            .attr('text-anchor', d => {
-                const angleDeg = (d.startAngle + d.endAngle) / 2 * 180 / Math.PI - 90;
-                return angleDeg > 90 ? 'end' : 'start';
-            })
-            .attr('opacity', d => {
-                const name = ordered[d.index].name;
-                return !selected || name === selected ? 1 : 0.35;
-            })
-            .attr('font-weight', d => ordered[d.index].name === selected ? 'bold' : 'normal')
-            // Text content is always populated; hidden labels are display:none
-            // so chord hover can unhide the two endpoint labels (and re-hide
-            // them on leave) without rewriting any DOM text.
-            .style('display', (d, i) => visible[i] ? null : 'none')
-            .text(d => this.showNames ? (labels.get(ordered[d.index].name) || ordered[d.index].name) : '');
-
-        // Chord hover surfaces the two endpoint labels even when they were
-        // hidden by the de-overlap pass — useful for tracing a ribbon back to
-        // both musicians. Namespaced so it doesn't trample _attachTooltip's
-        // own mouseenter/leave handlers on the same selection.
-        const labelTextSel = svg.select('g.network-arcs').selectAll('text.network-arc-label');
-        const showEndpoints = (event, d) => {
-            [d.source.index, d.target.index].forEach(idx => {
-                labelTextSel.filter(`[data-arc-index="${idx}"]`).style('display', null);
-            });
-        };
-        const restoreEndpoints = (event, d) => {
-            [d.source.index, d.target.index].forEach(idx => {
-                if (!visible[idx]) {
-                    labelTextSel.filter(`[data-arc-index="${idx}"]`).style('display', 'none');
-                }
-            });
-        };
-        chordSel
-            .on('mouseenter.labels', showEndpoints)
-            .on('mouseleave.labels', restoreEndpoints);
-    }
-
-    _truncate(text, maxWidthPx, fontPx) {
-        const charWidth = fontPx * 0.6;
-        const maxChars = Math.max(2, Math.floor(maxWidthPx / charWidth));
-        if (text.length <= maxChars) return text;
-        return text.slice(0, maxChars - 1) + '…';
     }
 
     // ---------------- Tooltip plumbing ----------------
